update query type names to better reflect operation; move from user resolver to contirbution resolver; update query
