+++ apps/protocol-api/src/prisma/resolvers/User.ts
@@ -2,8 +2,6 @@ import * as TypeGraphQL from 'type-graphql';
 
 import { Context } from './types';
 import { User } from '../generated/type-graphql/models/User';
-import { getPrismaFromContext } from '../generated/type-graphql/helpers';
-import { Int } from 'type-graphql';
 
 @TypeGraphQL.InputType('UserCreateCustomInput', {
   isAbstract: true,
@@ -35,18 +33,6 @@ export class UserUpdateCustomInput {
   disconnectLinearId?: number;
 }
 
-@TypeGraphQL.InputType('GetUserContributionCountInput')
-export class GetUserContributionCountInput {
-  @TypeGraphQL.Field((_type) => Number)
-  id: number;
-
-  @TypeGraphQL.Field((_type) => Date)
-  start_date: Date;
-
-  @TypeGraphQL.Field((_type) => Date)
-  end_date: Date;
-}
-
 @TypeGraphQL.ArgsType()
 export class UpdateUserCustomArgs {
   @TypeGraphQL.Field((_type) => UserUpdateCustomInput, {
@@ -73,14 +59,6 @@ export class GetUserArgs {
 export class ListUserArgs {
   @TypeGraphQL.Field((_type) => String)
   address: string;
-}
-
-@TypeGraphQL.ArgsType()
-export class GetUserContributionCountArgs {
-  @TypeGraphQL.Field((_type) => GetUserContributionCountInput, {
-    nullable: false,
-  })
-  where!: GetUserContributionCountInput;
 }
 
 @TypeGraphQL.Resolver((_of) => User)
@@ -118,6 +96,7 @@ export class UserCustomResolver {
       },
     });
   }
+
   @TypeGraphQL.Mutation((_returns) => User, { nullable: false })
   async updateUserCustom(
     @TypeGraphQL.Ctx() { prisma, req }: Context,
@@ -148,39 +127,4 @@ export class UserCustomResolver {
       where: { address: { equals: address } },
     });
   }
-  
-  @TypeGraphQL.Query((_returns) => Int, { nullable: false })
-  async getContributionCountForUser(
-    @TypeGraphQL.Ctx() ctx: Context,
-    @TypeGraphQL.Args() args: GetUserContributionCountArgs
-  ) {
-    /*return await getPrismaFromContext(ctx).contribution.count({
-      where: {
-        AND: [
-          { user_id: { equals: args.where.id } },
-          { date_of_engagement: { gte: args.where.start_date } },
-          { date_of_engagement: { lte: args.where.end_date } }
-        ]
-      }
-    })
-    // grouping by a derived field not yet supported in prisma
-    // would need to group by date, not the datetime as is stored in postg*/
-    // YYY-MM-DD hh:mm:ss.sss
-    const user_id  = args.where.id;
-    const start = fmt_date(args.where.start_date)
-    const end = fmt_date(args.where.end_date)
-    const query = `
-      SELECT date(date_of_engagement), count(date_of_engagement) as count
-      FROM "Contribution"
-      WHERE ("Contribution"."date_of_engagement" BETWEEN ${start} AND ${end}) AND
-      "Contribution".user_id = ${user_id} 
-      GROUP BY date(date_of_engagement)
-      ORDER BY date(date_of_engagement)
-    `;
-    return await getPrismaFromContext(ctx).$queryRaw`${query}`
-  }
-}
-
-function fmt_date(date: Date) {
-    return date.toISOString().replace(/[TZ]/g, " ").trim();
 }
